Move JavaInterface out from lib.rs to java.rs
